Moved validator and strip to their own files
