+++ main.py
@@ -31,6 +31,7 @@ def main():
     logging.getLogger("selenium").setLevel(logging.WARNING)
     logging.getLogger("PIL").setLevel(logging.WARNING)
     logging.getLogger("requests_oauthlib").setLevel(logging.WARNING)
+    logging.getLogger("chardet").setLevel(logging.WARNING)
 
     logging.info('Starting script')
